pyproject.toml
[project]
name = "retailworks-snowflake"
version = "0.1.0"
description = "Snowflake retail data warehouse project"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "snowflake-connector-python[pandas]",
    "snowflake-snowpark-python",
    "streamlit",
    "pandas",
    "numpy",
    "faker>=37.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
